+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -319,6 +319,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
     }
 
     @Test
+    @TestMetadata("privateConstructorSkipped.kt")
+    public void testPrivateConstructorSkipped() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/privateConstructorSkipped.kt");
+    }
+
+    @Test
     @TestMetadata("providerAndLazy.kt")
     public void testProviderAndLazy() {
       runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/providerAndLazy.kt");