+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/FastInitBoxTestGenerated.java
@@ -328,6 +328,12 @@ public class FastInitBoxTestGenerated extends AbstractFastInitBoxTest {
       }
 
       @Test
+      @TestMetadata("privateConstructorSkipped.kt")
+      public void testPrivateConstructorSkipped() {
+        runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/privateConstructorSkipped.kt");
+      }
+
+      @Test
       @TestMetadata("providerAndLazy.kt")
       public void testProviderAndLazy() {
         runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/providerAndLazy.kt");